Move elements under edit
