+++ components/form-builder/app/edit/Edit.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "next-i18next";
 
 import { LocalizedElementProperties, LocalizedFormProperties } from "../../types";
 import { useTemplateStore } from "../../store";
-import { RichTextLocked } from "../../elements";
+import { RichTextLocked } from "./elements";
 import { ElementPanel, ConfirmationDescription, PrivacyDescription } from ".";
 import { Input } from "../../shared";
