+++ components/form-builder/app/edit/elements/Option.tsx
@@ -1,9 +1,9 @@
 import React, { useRef, useEffect, ReactElement, useCallback, useState } from "react";
 import PropTypes from "prop-types";
-import { Close } from "../icons";
-import { Button } from "../shared/Button";
-import { Input } from "../shared/Input";
-import { useTemplateStore } from "../store/useTemplateStore";
+import { Close } from "../../../icons";
+import { Button } from "../../../shared/Button";
+import { Input } from "../../../shared/Input";
+import { useTemplateStore } from "../../../store/useTemplateStore";
 import { useTranslation } from "next-i18next";
 import debounce from "lodash.debounce";
